Thank you!
I wanted to give something back to your amazing library.  Here's some simplified code from the board I am working on (and nearly done with thanks to you). Feel free to use this as an example or change it to make it even more example worthy.  Any advice that will eliminate the need for that 'floatmap' function would be much appreciated.